+++ packages/elements/dropdown-menu/src/components/context/DropdownItemFocusManager.js
@@ -3,7 +3,7 @@
 import React, { Component, type Node } from 'react';
 import PropTypes from 'prop-types';
 import { focusManagerContext } from '../../util/contextNamespace';
-import type { ItemId, FocusItem } from '../../../src/types';
+import type { ItemId, FocusItem } from '../../types';
 import { KEY_DOWN, KEY_UP, KEY_TAB } from '../../util/keys';
 
 type Props = {