+++ packages/elements/icon/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/icon
 
+## 10.9.1
+- [patch] Packages Flow types for elements components [3111e74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3111e74)
+
 ## 10.9.0
 - [minor] New emoji-add icon [b8b1b51](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b8b1b51)