+++ packages/elements/modal-dialog/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/modal-dialog
 
+## 3.3.11
+- [patch] Packages Flow types for elements components [3111e74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3111e74)
+
 ## 3.3.10
 
 ## 3.3.9